Extract activity tab component rendering into a concern
The turbo-stream rendering helpers (index shell/list and journal item
show/edit) now live in ComponentStreaming, giving both the controller and
UpdateStreaming a single named source for item rendering instead of
UpdateStreaming reaching back into a controller private method.
